+++ app/src/main/java/com/example/task/ToDoFragment.kt
@@ -12,11 +12,14 @@ import android.widget.PopupWindow
 import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.activityViewModels
+import androidx.lifecycle.LiveData
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.task.adapters.PendingAdapter
+import com.example.task.database.Task
 import com.example.task.databinding.FragmentTodoBinding
 import com.example.task.models.*
+import com.google.android.material.snackbar.Snackbar
 import com.google.android.material.timepicker.MaterialTimePicker
 
 /**
@@ -31,11 +34,13 @@ class ToDoFragment : Fragment() {
 
     private var _binding: FragmentTodoBinding? = null
     private val binding get() = _binding!!
+    private lateinit var taskTodo: LiveData<List<Task>>
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View {
+        taskTodo = sharedViewModel.pendingTask
         _binding = FragmentTodoBinding.inflate(inflater, container, false)
         return binding.root
     }
@@ -57,7 +62,7 @@ class ToDoFragment : Fragment() {
             }
         }
 
-        sharedViewModel.pendingTask.observe(viewLifecycleOwner) {
+        sharedViewModel.pendingTask.observe(viewLifecycleOwner) { it ->
             binding.recyclerViewForFragmentToDo.adapter = PendingAdapter(it) { task, clickType ->
                 when (clickType) {
                     ClickType.SHORT -> sharedViewModel.markCompleted(task)
@@ -76,17 +81,22 @@ class ToDoFragment : Fragment() {
                 startActivity(intent)
             }
             binding.clearAll.setOnClickListener {
-                val alertDialog = AlertDialog.Builder(requireContext())
-                alertDialog.setMessage(R.string.delete_all_from_pending)
-                alertDialog.setCancelable(false)
-                alertDialog.setPositiveButton(android.R.string.yes) { _, _ ->
-                    sharedViewModel.deleteAllFromPending()
+                if(taskTodo.value!!.isEmpty()){
+                    Snackbar.make(it, R.string.list_empty,Snackbar.LENGTH_SHORT).show()
                 }
+                else{
+                    val alertDialog = AlertDialog.Builder(requireContext())
+                    alertDialog.setMessage(R.string.delete_all_from_pending)
+                    alertDialog.setCancelable(false)
+                    alertDialog.setPositiveButton(android.R.string.yes) { _, _ ->
+                        sharedViewModel.deleteAllFromPending()
+                    }
 
-                alertDialog.setNegativeButton(android.R.string.no) { dialog, _ ->
-                    dialog.cancel()
+                    alertDialog.setNegativeButton(android.R.string.no) { dialog, _ ->
+                        dialog.cancel()
+                    }
+                    alertDialog.show()
                 }
-                alertDialog.show()
             }
         }
     }
